+++ client-mod/thunderstore/changelog.md
@@ -7,6 +7,10 @@
 
 ## Unreleased ##
 
+### Changed ###
+
+* The two `ENABLE DEATH LINK`/`DISABLE DEATH LINK` buttons have been combined into a single one
+
 ### Fixed ###
 
 * Fixed an obscure bug which would cause the Whiplash to not work at all in the mission `1-4: COURT OF THE CORPSE KING`